backend/requirements.txt
pyqt6
llama-cpp-python
pyttsx3
sounddevice
numpy
